// src/app/model/procedurer-flat.ts
export class ProcedurerFlat {
    ProcedurerId: number;
    RegionId: number;
    RegionNamn: string;
    SourceGroupCode: string;
    SourceGroupDescription: string;
    ProcedurKod: string;
    ProcedurBeskrivning: string;
    ProcedurGiltig: boolean;
    ProcedurGiltigFrom: Date;
    ProcedurGiltigTo: Date;
    IsDoctor: boolean;
    Faktureras: string;
    AtenaNameing: string;
    DateCreated: Date;
    Provmaterial: string;
    RegelTypeId: number;
    RegelTypeName: string;

   }